created a new algorithm for a non folding mesh.
I discovered a new algorithm which may be able to minimize the area functional. take a non folding mesh on the grid with nx by ny points, use lloyd's algorithm to get those points to have a voronoi diagram with equal areas, and approximate the unstructured voronoi mesh with a structured one.